+++ packages/react/src/cart-hooks.tsx
@@ -8,11 +8,13 @@ import {
   SHOPIFY_STOREFRONT_ID_HEADER,
   SHOPIFY_STOREFRONT_Y_HEADER,
   SHOPIFY_STOREFRONT_S_HEADER,
+} from './cart-constants.js';
+import {
   SHOPIFY_Y,
   SHOPIFY_S,
-} from './cart-constants.js';
-import {parse} from 'worktop/cookie';
+} from './shared-constants.js';
 import type {StorefrontApiResponseOkPartial} from './storefront-api-response.types.js';
+import { getShopifyCookies } from './cookies-utils.js';
 
 export function useCartFetch() {
   const {storefrontId, getPublicTokenHeaders, getStorefrontApiUrl} = useShop();
@@ -32,11 +34,9 @@ export function useCartFetch() {
       }
 
       // Find Shopify cookies
-      const cookieData = parse(document.cookie);
-      if (cookieData[SHOPIFY_Y] && cookieData[SHOPIFY_S]) {
-        headers[SHOPIFY_STOREFRONT_Y_HEADER] = cookieData[SHOPIFY_Y];
-        headers[SHOPIFY_STOREFRONT_S_HEADER] = cookieData[SHOPIFY_S];
-      }
+      const cookieData = getShopifyCookies(document.cookie);
+      headers[SHOPIFY_STOREFRONT_Y_HEADER] = cookieData[SHOPIFY_Y];
+      headers[SHOPIFY_STOREFRONT_S_HEADER] = cookieData[SHOPIFY_S];
 
       return fetch(getStorefrontApiUrl(), {
         method: 'POST',
