+++ packages/react/src/shared-constants.ts
@@ -1,0 +1,2 @@
+export const SHOPIFY_Y = '_shopify_y';
+export const SHOPIFY_S = '_shopify_s';
